fix(ubuntu): force restart check only on WSL 2

+++ ellipsis.sh
@@ -7,12 +7,16 @@ pkg.link() {
 
 pkg.install() {
     [ -f "$PKG_PATH/install.sh" ] && bash $PKG_PATH/install.sh "$ELLIPSIS_SRC" "$PKG_PATH"
+    
+    KERNEL_VERSION=`cat /proc/version`
 
-    [ -f ".restart.lock" ] &&
-      echo "" &&
-      echo -e "\e[33mPlease restart the computer and then re-run the ellipsis command from a WSL prompt to continue the installation.\e[0m" &&
-      rm -rf .restart.lock &&
-      exit 1
+    if [[ "$KERNEL_VERSION" == *"microsoft"* ]]; then
+        [ -f ".restart.lock" ] &&
+        echo "" &&
+        echo -e "\e[33mPlease restart the computer and then re-run the ellipsis command from a WSL prompt to continue the installation.\e[0m" &&
+        rm -rf .restart.lock &&
+        exit 1
+    fi
 }
 
 pkg.pull() {
